config: validate configuration before operating on it
This patch implements some refactoring to separate configuration parsing
and validation into its own compile unit.  It also adds some basic
validation of the parsed configuration.
