+++ app/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(magenta magenta.cc)
+add_executable(magenta cmd.cc magenta.cc)
 
 target_include_directories(magenta PRIVATE
   ${PROJECT_SOURCE_DIR}/external/args
